+++ examples/build-and-push-join-context/build-and-push-join-context.go
@@ -79,7 +79,8 @@ func buildAndPushJoinContext(w io.Writer) error {
 		response.WithWriter(w),
 	)
 
-	dockerBuilder := build.NewDockerBuildCmd(dockerCli, imageName).
+	dockerBuilder := build.NewDockerBuildCmd(dockerCli).
+		WithImageName(imageName).
 		WithPushAfterBuild().
 		WithRemoveAfterPush().
 		WithResponse(res)