+++ examples/build-and-push/build-and-push.go
@@ -52,7 +52,8 @@ func buildAndPush(w io.Writer) error {
 		response.WithWriter(w),
 	)
 
-	dockerBuilder := build.NewDockerBuildCmd(dockerCli, imageName).
+	dockerBuilder := build.NewDockerBuildCmd(dockerCli).
+		WithImageName(imageName).
 		WithPushAfterBuild().
 		WithRemoveAfterPush().
 		WithResponse(res)