+++ examples/build-git-context/build-git-context.go
@@ -43,7 +43,8 @@ func buildGitContext(w io.Writer) error {
 		response.WithWriter(w),
 	)
 
-	dockerBuilder := build.NewDockerBuildCmd(dockerCli, imageName).
+	dockerBuilder := build.NewDockerBuildCmd(dockerCli).
+		WithImageName(imageName).
 		WithPushAfterBuild().
 		WithRemoveAfterPush().
 		WithResponse(res)